+++ src/vite-env.d.ts
@@ -1,7 +1,9 @@
 /// <reference types="vite/client" />
 /// <reference types="vite-plugin-pwa/client" />
 
-
+declare const __APP_VERSION__: string;
+declare const __APP_BUILD_COMMIT__: string;
+declare const __APP_BUILD_DATE__: string;
 
 interface ImportMeta {
     readonly env: ImportMetaEnv;